+++ Emulator/Components/Emulator.h
@@ -106,20 +106,20 @@ public:
 
     // Queries an option
     i64 get(Option option) const;
-    i64 get(Option option, long id) const;
+    i64 get(Option option, isize id) const;
 
     // Configures the emulator to match a specific C64 model
     void set(C64Model model);
 
     // Sets an option
     void set(Option option, i64 value) throws;
-    void set(Option option, long id, i64 value) throws;
+    void set(Option option, isize id, i64 value) throws;
 
     // Convenience wrappers
     void set(Option opt, const string &value) throws;
-    void set(Option opt, long id, const string &value) throws;
+    void set(Option opt, isize id, const string &value) throws;
     void set(const string &opt, const string &value) throws;
-    void set(const string &opt, long id, const string &value) throws;
+    void set(const string &opt, isize id, const string &value) throws;
 
     // Returns the emulated refresh rate of the virtual C64
     double refreshRate() const override;
